// src/modules/todo-list/enums/api-task-error.enums.ts
export enum ApiTaskErrorEnums {
    FAILED_TO_CREATE_TASK = 'Failed to create task',
    FAILED_TO_GET_TASKS = 'Failed to get user tasks',
    FAILED_TO_DELETE_TASK = 'Failed to delete task',
    FAILED_TO_GET_IMPORTANT_TASKS = 'Failed to get important user tasks',
    FAILED_TO_GET_TODAY_TASKS = 'Failed to get today user tasks',
    FAILED_TO_GET_USER_TASKS = 'Failed to get today user tasks for a specific date',
    FAILED_TO_GET_TASK_BY_ID = 'Failed to get task by id',
    FAILED_TO_UPDATE_TASK = 'Failed to update task',
}